backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
openai==1.30.0
python-dotenv==1.0.1
google-search-results==2.4.2
